refactor: move duplicate code fragments for examples to util package
