+++ examples/express-s3/express-s3.ts
@@ -1,10 +1,11 @@
-import baggyRouter, { ResponseDecorator } from "@baggy/express-router";
+import baggyRouter from "@baggy/express-router";
 import express from "express";
 import ProviderS3 from "@baggy/provider-s3";
 import { Registry } from "@baggy/registry";
-import { errorHandler, requestLogger, debug } from "./logger";
+import { errorHandler, requestLogger } from "../util/logger";
 import chalk from "chalk";
 import S3 from "aws-sdk/clients/s3";
+import { responseDecorator } from "../util/response-decorator";
 
 const app = express();
 const basePath = ".local";
@@ -35,17 +36,8 @@ function info(...str: string[]): void {
   console.log(chalk.green.bold("Baggy Express-S3"), chalk.cyan(...str));
 }
 
-const decorator: ResponseDecorator = (result, request) => {
-  let { statusCode, body } = result; // eslint-disable-line prefer-const
-  if (body instanceof Buffer) {
-    body = body.toString();
-  }
-  debug(Date.now() + `${request.path.replace(/\//g, "_")}.json`, JSON.stringify({ statusCode, body }, null, 2), true);
-  return result;
-};
-
 app.use(requestLogger(requestLogs));
-app.use(baggyRouter(registry, decorator));
+app.use(baggyRouter(registry, responseDecorator));
 app.use(errorHandler(errorLog));
 app.listen(3000, () => {
   info("Starting server on http://localhost:3000");
