+++ examples/express-sqlite/express-sqlite.ts
@@ -1,11 +1,12 @@
-import router, { ResponseDecorator } from "@baggy/express-router";
+import router from "@baggy/express-router";
 import express from "express";
 import sqlite3 from "sqlite3";
 import Database from "@baggy/provider-sqlite";
 import fs from "fs";
 import { Registry } from "@baggy/registry";
-import { errorHandler, requestLogger, debug } from "./logger";
+import { errorHandler, requestLogger } from "../util/logger";
 import chalk from "chalk";
+import { responseDecorator } from "../util/response-decorator";
 
 const app = express();
 const basePath = ".local";
@@ -51,17 +52,8 @@ const registry = new Registry({
   proxyCache: true,
 });
 
-const decorator: ResponseDecorator = (result, request) => {
-  let { statusCode, body } = result; // eslint-disable-line prefer-const
-  if (body instanceof Buffer) {
-    body = body.toString();
-  }
-  debug(Date.now() + `${request.path.replace(/\//g, "_")}.json`, JSON.stringify({ statusCode, body }, null, 2), true);
-  return result;
-};
-
 app.use(requestLogger(requestLogs));
-app.use(router(registry, decorator));
+app.use(router(registry, responseDecorator));
 app.use(errorHandler(errorLog));
 app.listen(3000, () => {
   info("Starting server on http://localhost:3000");
